+++ src/reducers/controls.ts
@@ -156,11 +156,12 @@ export interface MeasurementFilters {
   [key: string]: Map<string, {active: boolean}>
 }
 export interface MeasurementsControlState {
-  measurementsGroupBy: string | undefined,
-  measurementsDisplay: MeasurementsDisplay | undefined,
-  measurementsShowOverallMean: boolean | undefined,
-  measurementsShowThreshold: boolean | undefined,
+  measurementsGroupBy: string | undefined
+  measurementsDisplay: MeasurementsDisplay | undefined
+  measurementsShowOverallMean: boolean | undefined
+  measurementsShowThreshold: boolean | undefined
   measurementsFilters: MeasurementFilters
+  measurementsColorGrouping: string | undefined
 }
 
 export interface ControlsState extends BasicControlsState, MeasurementsControlState {}
@@ -247,6 +248,7 @@ export const getDefaultControlsState = (): ControlsState => {
     measurementsDisplay: undefined,
     measurementsShowOverallMean: undefined,
     measurementsShowThreshold: undefined,
+    measurementsColorGrouping: undefined,
     measurementsFilters: {},
     performanceFlags: new Map(),
   };
@@ -266,7 +268,8 @@ export const defaultMeasurementsControlState: MeasurementsControlState = {
   measurementsDisplay: "mean",
   measurementsShowOverallMean: true,
   measurementsShowThreshold: true,
-  measurementsFilters: {}
+  measurementsFilters: {},
+  measurementsColorGrouping: undefined,
 };
 
 /* while this may change, div currently doesn't have CIs, so they shouldn't be displayed. */
@@ -479,6 +482,13 @@ const Controls = (state: ControlsState = getDefaultControlsState(), action): Con
       }
       return newState;
     }
+    case types.REMOVE_METADATA: {
+      const coloringsPresentOnTree = new Set(state.coloringsPresentOnTree);
+      action.nodeAttrsToRemove.forEach((colorBy: string): void => {
+        coloringsPresentOnTree.delete(colorBy);
+      })
+      return {...state, coloringsPresentOnTree};
+    }
     case types.UPDATE_VISIBILITY_AND_BRANCH_THICKNESS: {
       const colorScale = Object.assign({}, state.colorScale, { visibleLegendValues: action.visibleLegendValues });
       return Object.assign({}, state, { colorScale: colorScale });
@@ -495,6 +505,7 @@ const Controls = (state: ControlsState = getDefaultControlsState(), action): Con
       return state;
     }
     case types.CHANGE_MEASUREMENTS_COLLECTION: // fallthrough
+    case types.CHANGE_MEASUREMENTS_COLOR_GROUPING: // fallthrough
     case types.CHANGE_MEASUREMENTS_DISPLAY: // fallthrough
     case types.CHANGE_MEASUREMENTS_GROUP_BY: // fallthrough
     case types.TOGGLE_MEASUREMENTS_OVERALL_MEAN: // fallthrough